Added betroll bettest, fixed a bug which caused betroll to have very low payout

+++ src/Wiz.Econ/Gambling/Betflip/BetflipGame.cs
@@ -12,7 +12,7 @@ public sealed class BetflipGame
 
     public BetflipResult Flip(byte guess, decimal amount)
     {
-        var side = _rng.Next(0, 2);
+        var side = (byte)_rng.Next(0, 2);
         if (side == guess)
         {
             return new BetflipResult()
